+++ agent-inject/agent/container_sidecar.go
@@ -18,6 +18,7 @@ const (
 	DefaultContainerArg       = "echo ${VAULT_CONFIG?} | base64 -d > /home/vault/config.json && vault agent -config=/home/vault/config.json"
 	DefaultRevokeGrace        = 5
 	DefaultAgentLogLevel      = "info"
+	DefaultAgentLogFormat     = "standard"
 )
 
 // ContainerSidecar creates a new container to be added
@@ -43,6 +44,10 @@ func (a *Agent) ContainerSidecar() (corev1.Container, error) {
 			MountPath: extraSecretVolumePath,
 			ReadOnly:  true,
 		})
+	}
+
+	if a.CopyVolumeMounts != "" {
+		volumeMounts = append(volumeMounts, a.copyVolumeMounts(a.CopyVolumeMounts)...)
 	}
 
 	arg := DefaultContainerArg